Added folders to src
